requirements.txt
flask
pandas
requests
azure.storage
azure.cosmosdb.table